+++ docs/NEWS.rst
@@ -8,6 +8,13 @@ of XIST. For a description of how to update your code to each versions of XIST
 see :ref:`MIGRATION`.
 
 
+Changes in 5.69.1 (released 12/13/2021)
+---------------------------------------
+
+*	Fixed a bug in the PySQL command :class:`~ll.pysql.commit`: Removed the
+	useless argument ``sql``.
+
+
 Changes in 5.69 (released 11/17/2021)
 -------------------------------------